# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Task Manager API
  description: |
    Backend API for the Task Manager application.
    Authentication is handled on the frontend (Google, GitHub, Microsoft, LinkedIn OAuth).
    The backend expects a valid JWT Bearer token in the Authorization header.
  version: 1.0.0
  contact:
    name: Task Manager Team
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /api/users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [Users]
      summary: Update current user profile
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/users:
    get:
      tags: [Users]
      summary: List / search users
      operationId: listUsers
      parameters:
        - name: q
          in: query
          description: Search by name or email
          schema:
            type: string
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /api/users/invite:
    post:
      tags: [Users]
      summary: Invite a user by email
      operationId: inviteUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUserRequest"
      responses:
        "201":
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Projects ───────────────────────────────────────────
...
  /api/projects/{projectId}/tasks:
    get:
      tags: [Tasks]
      summary: List tasks with filters
      operationId: listTasks
      parameters:
        - $ref: "#/components/parameters/ProjectIdParam"
        - name: status
          in: query
          description: "Filter by status (comma-separated)"
          schema:
            type: string
            enum: [todo, in_progress, staging_review, done]
        - name: priority
          in: query
          description: "Filter by priority"
          schema:
            type: string
            enum: [p1, p2, p3, p4]
        - name: assignee_id
          in: query
          description: Filter by assignee user ID
          schema:
            type: string
        - name: reporter_id
          in: query
          description: Filter by reporter user ID
          schema:
            type: string
        - name: label_ids
          in: query
          description: "Filter by label IDs (comma-separated)"
          schema:
            type: string
        - name: due_before
          in: query
          description: Tasks due before this date (RFC 3339)
          schema:
            type: string
            format: date-time
        - name: due_after
          in: query
          description: Tasks due after this date (RFC 3339)
          schema:
            type: string
            format: date-time
        - name: is_past_due
          in: query
          description: Filter tasks past their due date
          schema:
            type: boolean
        - name: sprint_id
          in: query
          description: Filter by sprint ID
          schema:
            type: string
        - name: sort_by
          in: query
          description: "Sort field"
          schema:
            type: string
            enum: [due_date, priority, created_at, updated_at]
            default: created_at
        - name: sort_order
          in: query
          description: "Sort direction"
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
    post:
      tags: [Tasks]
      summary: Create a new task
      operationId: createTask
      parameters:
        - $ref: "#/components/parameters/ProjectIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: Get task by ID
      operationId: getTask
      parameters:
        - $ref: "#/components/parameters/TaskIdParam"
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Tasks]
      summary: Update a task
      operationId: updateTask
      parameters:
        - $ref: "#/components/parameters/TaskIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Tasks]
      summary: Delete a task
      operationId: deleteTask
      parameters:
        - $ref: "#/components/parameters/TaskIdParam"
      responses:
        "204":
          description: Task deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/tasks/{taskId}/assign:
    put:
      tags: [Tasks]
      summary: Assign task to a user
      operationId: assignTask
      parameters:
        - $ref: "#/components/parameters/TaskIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTaskRequest"
      responses:
        "200":
          description: Task assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/tasks/{taskId}/status:
    put:
      tags: [Tasks]
      summary: Update task status
      operationId: updateTaskStatus
      parameters:
        - $ref: "#/components/parameters/TaskIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/tasks/{taskId}/time:
    get:
      tags: [Tasks]
      summary: Get time tracking info for a task
      operationId: getTaskTimeTracking
      parameters:
        - $ref: "#/components/parameters/TaskIdParam"
      responses:
        "200":
          description: Time tracking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeTracking"
    put:
      tags: [Tasks]
      summary: Log time against a task
      operationId: logTaskTime
      parameters:
        - $ref: "#/components/parameters/TaskIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogTimeRequest"
      responses:
        "200":
          description: Time logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeTracking"
  /api/tasks/my:
    get:
      tags: [Tasks]
      summary: Get tasks assigned to the current user across all projects
      operationId: getMyTasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, staging_review, done]
        - name: is_past_due
          in: query
          schema:
            type: boolean
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Current user's tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"

  # ─── Comments ──────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from frontend OAuth flow
  parameters:
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      schema:
        type: string
    TaskIdParam:
      name: taskId
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          example: "usr_abc123"
        email:
          type: string
          format: email
          example: "ayush@example.com"
        name:
          type: string
          example: "Ayush"
        avatar_url:
          type: string
          format: uri
        password:
          type: string
          format: password
          writeOnly: true
          example: "secret123"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        avatar_url:
          type: string
          format: uri
    InviteUserRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        project_id:
          type: string
          description: Optionally invite directly to a project
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # ── Project ───────────────────────────
...
    Task:
      type: object
      properties:
        id:
          type: string
          example: "tsk_001"
        project_id:
          type: string
        title:
          type: string
          example: "Implement login flow"
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, staging_review, done]
          example: "todo"
        priority:
          type: string
          enum: [p1, p2, p3, p4]
          description: "P1=Critical, P2=High, P3=Medium, P4=Low"
          example: "p2"
        due_date:
          type: string
          format: date-time
          nullable: true
        is_past_due:
          type: boolean
          description: "Computed: true if due_date < now and status != done"
        assignee_id:
          type: string
          nullable: true
        assignee:
          $ref: "#/components/schemas/User"
        reporter_id:
          type: string
        reporter:
          $ref: "#/components/schemas/User"
        label_ids:
          type: array
          items:
            type: string
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
        sprint_id:
          type: string
          nullable: true
        comment_count:
          type: integer
        time_tracking:
          $ref: "#/components/schemas/TimeTracking"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          maxLength: 10000
        status:
          type: string
          enum: [todo, in_progress, staging_review, done]
          default: todo
        priority:
          type: string
          enum: [p1, p2, p3, p4]
          default: p3
        due_date:
          type: string
          format: date-time
        assignee_id:
          type: string
        label_ids:
          type: array
          items:
            type: string
        sprint_id:
          type: string
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          maxLength: 10000
        status:
          type: string
          enum: [todo, in_progress, staging_review, done]
        priority:
          type: string
          enum: [p1, p2, p3, p4]
        due_date:
          type: string
          format: date-time
          nullable: true
        assignee_id:
          type: string
          nullable: true
        label_ids:
          type: array
          items:
            type: string
        sprint_id:
          type: string
          nullable: true
    AssignTaskRequest:
      type: object
      required: [assignee_id]
      properties:
        assignee_id:
          type: string
    UpdateStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [todo, in_progress, staging_review, done]
    TaskListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        pagination:
          $ref: "#/components/schemas/Pagination"
    TimeTracking:
      type: object
      properties:
        estimated_hours:
          type: number
          format: float
        logged_hours:
          type: number
          format: float
        is_overdue:
          type: boolean
          description: "True if due_date has passed and task is not done"
        overdue_duration:
          type: string
          description: "Human-readable overdue duration, e.g. '2d 3h'"
          nullable: true
    LogTimeRequest:
      type: object
      required: [hours]
      properties:
        hours:
          type: number
          format: float
          minimum: 0.01
        description:
          type: string
          maxLength: 500

    # ── Comment ───────────────────────────
...
    Label:
      type: object
      properties:
        id:
          type: string
          example: "lbl_001"
        project_id:
          type: string
        name:
          type: string
          example: "bug"
        color:
          type: string
          description: "Hex color code"
          example: "#FF5733"
        created_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    MessageResponse:
      type: object
      properties:
        message:
          type: string